fix: sanaei add panel  process

+++ database/migrations/2024_09_15_145936_create_cron_logs_table.php
@@ -4,23 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('cron_logs', function (Blueprint $table) {
-            $table->id();
-            $table->bigInteger('cron_id')->unsigned();
-            $table->bigInteger('product_id')->unsigned();
-            $table->timestamps();
-            $table->foreign('cron_id')->references('id')->on('cron_jobs')->onDelete('cascade');
-            $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-
-
-        });
+        if (!Schema::hasTable('cron_logs')) {
+            Schema::create('cron_logs', function (Blueprint $table) {
+                $table->id();
+                $table->bigInteger('cron_id')->unsigned();
+                $table->bigInteger('product_id')->unsigned();
+                $table->timestamps();
+                $table->foreign('cron_id')->references('id')->on('cron_jobs')->onDelete('cascade');
+                $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
+            });
+        }
     }
 
     /**
